Adding condition on model training upon start up

+++ domaine/ml_model.py
@@ -5,6 +5,7 @@ from sklearn.model_selection import train_test_split
 from sklearn.preprocessing import StandardScaler
 from sklearn.metrics import accuracy_score
 import pickle   
+import os
 from csv import writer
 
 
@@ -14,7 +15,7 @@ def add_data(wine):
     Args:
         wine (WineModel): the wine we want to add
     """
-    with open('event.csv', 'a') as f_object:
+    with open('/home/yasait/Wine-Quality-Prediction/datasource/Wines.csv', 'a') as f_object:
         writer_object = writer(f_object)
         writer_object.writerow(wine)
         f_object.close()
@@ -40,18 +41,19 @@ def get_model_info():
     #to do 
     return info
 
-
-
-df= pd.read_csv("/home/yasait/Wine-Quality-Prediction/datasource/Wines.csv")
-X = df.drop(columns=['quality'],axis=1)
-y = df["quality"]
-X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.3, random_state = 0)
-sc_x = StandardScaler()
-X_train = sc_x.fit_transform(X_train)
-X_test = sc_x.fit_transform(X_test)
-dTree_clf = DecisionTreeClassifier()
-dTree_clf.fit(X_train,y_train)
-filename = 'finalized_model.sav'
-pickle.dump(dTree_clf, open(filename, 'wb'))
-y_pred2 = dTree_clf.predict(X_test)
-print("Accuracy of Model1::",accuracy_score(y_test,y_pred2))
+path="'domaine/finalized_model.pkl'"
+#Train the model only if it's the first lunch
+if os.path.exists(path)== False :
+    df= pd.read_csv("/home/yasait/Wine-Quality-Prediction/datasource/Wines.csv")
+    X = df.drop(columns=['quality'],axis=1)
+    y = df["quality"]
+    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.3, random_state = 0)
+    sc_x = StandardScaler()
+    X_train = sc_x.fit_transform(X_train)
+    X_test = sc_x.fit_transform(X_test)
+    dTree_clf = DecisionTreeClassifier()
+    dTree_clf.fit(X_train,y_train)
+    filename = 'domaine/finalized_model.pkl'
+    pickle.dump(dTree_clf, open(filename, 'wb'))
+    y_pred2 = dTree_clf.predict(X_test)
+    print("Accuracy of Model1::",accuracy_score(y_test,y_pred2))
